Split user_commands.py by command
For consistency with channel commands.
